+++ src/commands/music/stop.ts
@@ -4,7 +4,7 @@ export const command: ICommand = {
   name: "stop",
   aliases: [],
   requireOnSpecificChannel: true,
-  run: async (client, message, args) => {
+  run: async (client, message, args, noRemoveMessage) => {
     if (!message.member.voice.channel)
       return message.channel
         .send(`You're not in a voice channel !`)
@@ -35,6 +35,10 @@ export const command: ICommand = {
     message.channel.send(`Music **stopped** into this server !`).then((msg) => {
       setTimeout(() => msg.delete(), 5000);
     });
+
+    if (noRemoveMessage) {
+      return;
+    }
 
     message.delete();
   },
